sim search interface and services
