+++ backend/src/app/core/db/index_type.py
@@ -1,7 +1,7 @@
-from enum import Enum
+from enum import StrEnum
 
 
-class IndexType(str, Enum):
+class IndexType(StrEnum):
     SENTENCE = "sentence"
     IMAGE = "image"
     NAMED_ENTITY = "named-entity"
