+++ backend/src/app/core/search/typesense_service.py
@@ -4,33 +4,37 @@ import numpy as np
 import typesense
 from loguru import logger
 
-from app.core.data.dto.search import SimSearchSentenceHit
-from app.preprocessing.ray_model_service import RayModelService
-from app.preprocessing.ray_model_worker.dto.clip import ClipTextEmbeddingInput
-from app.util.singleton_meta import SingletonMeta
+from app.core.data.dto.search import SimSearchImageHit, SimSearchSentenceHit
+from app.core.search.index_type import IndexType
+from app.core.search.vector_index_service import VectorIndexService
 from config import conf
 
 
-class TypesenseService(metaclass=SingletonMeta):
+class TypesenseService(VectorIndexService):
     def __new__(cls, *args, **kwargs):
         cls._sentence_class_name = "Sentence"
         cls._image_class_name = "Image"
         cls._named_entity_class_name = "NamedEntity"
         cls._document_class_name = "Document"
 
+        cls.class_names = {
+            IndexType.SENTENCE: cls._sentence_class_name,
+            IndexType.IMAGE: cls._image_class_name,
+            IndexType.NAMED_ENTITY: cls._named_entity_class_name,
+            IndexType.DOCUMENT: cls._document_class_name,
+        }
+
         cls._common_fields = [
             {"name": "vec", "type": "float[]", "num_dim": 512},
             {"name": "project_id", "type": "int32"},
             {"name": "sdoc_id", "type": "int32"},
             {"name": "text", "type": "string"},
+            {"name": "sentence_id", "type": "int32"},
         ]
 
         cls._sentence_schema = {
             "name": cls._sentence_class_name,
-            "fields": [
-                *cls._common_fields,
-                {"name": "sentence_id", "type": "int32"},
-            ],
+            "fields": [*cls._common_fields],
         }
 
         cls._document_schema = {
@@ -38,9 +42,15 @@ class TypesenseService(metaclass=SingletonMeta):
             "fields": cls._common_fields,
         }
 
+        cls._image_schema = {
+            "name": cls._image_class_name,
+            "fields": cls._common_fields,
+        }
+
         cls._colletions = {
             cls._sentence_class_name: cls._sentence_schema,
             cls._document_class_name: cls._document_schema,
+            cls._image_class_name: cls._image_schema,
         }
 
         try:
@@ -70,65 +80,43 @@ class TypesenseService(metaclass=SingletonMeta):
             logger.error(msg)
             raise SystemExit(msg)
 
-        cls.rms = RayModelService()
-
         return super(TypesenseService, cls).__new__(cls)
 
-    def add_text_sdoc_to_index(
-        self,
-        proj_id: int,
-        sdoc_id: int,
-        sentences: List[str],
-    ) -> None:
-        sentence_embs = self.rms.clip_text_embedding(
-            ClipTextEmbeddingInput(text=sentences)
-        ).numpy()
-
-        # create cheap&easy (but suboptimal) document embeddings for now
-        doc_emb = sentence_embs.sum(axis=0)
-        doc_emb /= np.linalg.norm(doc_emb)
-
+    def add_embeddings_to_index(
+        self, type: IndexType, proj_id: int, sdoc_id: int, embeddings: List[np.ndarray]
+    ):
+        collection_name = self.class_names[type]
         logger.debug(
-            f"Adding {len(sentence_embs)} sentences "
+            f"Adding {len(embeddings)} embeddeddings "
             f"from SDoc {sdoc_id} in Project {proj_id} to Typesense ..."
         )
         sents = [
             {
-                "id": f"{proj_id}-{sdoc_id}-{sent_id}",
+                "id": (
+                    f"{sdoc_id}-{sent_id}"
+                    if type == IndexType.SENTENCE
+                    else str(sdoc_id)
+                ),
                 "project_id": proj_id,
                 "sdoc_id": sdoc_id,
                 "sentence_id": sent_id,
-                "text": sentences[sent_id],
                 "vec": sent_emb.tolist(),
             }
-            for sent_id, sent_emb in enumerate(sentence_embs)
+            for sent_id, sent_emb in enumerate(embeddings)
         ]
-        res = self._client.collections[self._sentence_class_name].documents.import_(
+        res = self._client.collections[collection_name].documents.import_(
             sents, {"action": "create"}
         )
         print(res)
         print("added sentences to TS", len(sents))
-        documents = [
-            {
-                "id": f"{proj_id}-{sdoc_id}",
-                "project_id": proj_id,
-                "sdoc_id": sdoc_id,
-                "text": " ".join(sentences),
-                "vec": doc_emb.tolist(),
-            }
-        ]
-        self._client.collections[self._document_class_name].documents.import_(
-            documents, {"action": "create"}
+
+    def remove_embeddings_from_index(self, type: IndexType, sdoc_id: int):
+        logger.debug(f"Removing text SDoc {sdoc_id} from Index!")
+        self._client.collections[self.class_names[type]].documents.delete(
+            {"filter_by": f"sdoc_id:={sdoc_id}"}
         )
 
-    def remove_text_sdoc_from_index(self, sdoc_id: int) -> None:
-        logger.debug(f"Removing text SDoc {sdoc_id} from Index!")
-        for name in self._colletions.keys():
-            self._client.collections[name].documents.delete(
-                {"filter_by": f"sdoc_id:={sdoc_id}"}
-            )
-
-    def remove_all_project_embeddings(
+    def remove_project_from_index(
         self,
         proj_id: int,
     ) -> None:
@@ -137,22 +125,54 @@ class TypesenseService(metaclass=SingletonMeta):
                 {"filter_by": f"project_id:={proj_id}"}
             )
 
-    def suggest_similar_sentences(
-        self, proj_id: int, sdoc_sent_ids: List[Tuple[int, int]]
-    ) -> List[SimSearchSentenceHit]:
-        return self.__suggest(proj_id, sdoc_sent_ids)
-
-    def __suggest(
+    def search_index(
         self,
         proj_id: int,
-        sdoc_sent_ids: List[Tuple[int, int]],
+        index_type: IndexType,
+        query_emb: np.ndarray,
+        sdoc_ids_to_search: List[int],
         top_k: int = 10,
         threshold: float = 0.0,
+    ) -> List[SimSearchSentenceHit] | List[SimSearchImageHit]:
+
+        results = self._client.collections[
+            self.class_names[index_type]
+        ].documents.search(
+            {
+                "vector_query": f"vec:({query_emb.tolist()}, k:{top_k})",
+                "filter_by": f"project_id:= {proj_id}",
+                "include_fields": "id,sdoc_id,sentence_id",
+            }
+        )
+
+        if index_type == IndexType.SENTENCE:
+            return [
+                SimSearchSentenceHit(
+                    sdoc_id=r["document"]["sdoc_id"],
+                    sentence_id=r["document"]["sentence_id"],
+                    score=r["vector_distance"],
+                )
+                for r in results
+            ]
+        else:
+            return [
+                SimSearchImageHit(
+                    sdoc_id=r["document"]["sdoc_id"],
+                    score=r["vector_distance"],
+                )
+                for r in results
+            ]
+
+    def suggest(
+        self,
+        index_type: IndexType,
+        proj_id: int,
+        sdoc_sent_ids: List[Tuple[int, int]],
     ) -> List[SimSearchSentenceHit]:
         candidates: List[SimSearchSentenceHit] = []
         vc = "vector_query"
         queries = [
-            {vc: f"vec:([], id: {proj_id}-{sdoc_id}-{sent_id}, k:1)"}
+            {vc: f"vec:([], id: {sdoc_id}-{sent_id}, k:1)"}
             for sdoc_id, sent_id in sdoc_sent_ids
         ]
 
@@ -176,16 +196,4 @@ class TypesenseService(metaclass=SingletonMeta):
                     )
                 )
 
-        candidates.sort(key=lambda x: (x.sdoc_id, x.sentence_id))
-        hits = self.__unique_consecutive(candidates)
-        hits = [h for h in hits if (h.sdoc_id, h.sentence_id) not in sdoc_sent_ids]
-        return hits
-
-    def __unique_consecutive(self, hits: List[SimSearchSentenceHit]):
-        result = []
-        current = SimSearchSentenceHit(sdoc_id=-1, sentence_id=-1, score=0.0)
-        for hit in hits:
-            if hit.sdoc_id != current.sdoc_id or hit.sentence_id != current.sentence_id:
-                current = hit
-                result.append(hit)
-        return result
+        return candidates
