+++ docker/.env.example
@@ -9,7 +9,7 @@ COMPOSE_PROJECT_NAME=demo
 # dependencies are as follows:
 # ray < background < backend < frontend
 # you can remove profiles from right to left
-COMPOSE_PROFILES=ray,background,backend,frontend
+COMPOSE_PROFILES=ray,background,backend,frontend,weaviate
 # Docker tag to use for pulling the backend and celery containers
 DATS_BACKEND_DOCKER_VERSION=1.1.1
 # Docker tag to use for pulling the ray container
@@ -39,6 +39,11 @@ JWT_SECRET=
 # The system user is automatically created and owns automatically generated data.
 SYSTEM_USER_EMAIL="SYSTEM@dats.org"
 SYSTEM_USER_PASSWORD="12SYSTEM34"
+
+# choose vector index service: weaviate, typesense, qdrant
+# enable the corresponding COMPOSE_PROFILES as well
+VECTOR_INDEX=weaviate
+
 # Mail settings
 MAIL_ENABLED=False
 MAIL_FROM=dats@uni-hamburg.de
